optimization: delay(uint16_t t)

+++ src/GTPowerA610.cpp
@@ -90,7 +90,7 @@ void hardware::setFan(bool enable)
 }
 
 
-void hardware::delay(uint32_t t)
+void hardware::delay(uint16_t t)
 {
 //	timer.delay(t);
 	analogInputs.doMeasurement((AnalogInputs::PHYSICAL_INPUTS*t)/4);
